+++ ports/Arduino/Arduino-generic/api/pin/write.c.v
@@ -6,6 +6,6 @@ module pin
 
 // write puts a logic value to a specific pin
 @[inline]
-pub fn write(name int, val int) {  
+pub fn write(name u8, val u8) {  
     C.digitalWrite(name, val)
 }
